fix dashboard

+++ database/migrations/2021_08_16_144354_add_payment_condition_to_invoices.php
@@ -14,7 +14,7 @@ class AddPaymentConditionToInvoices extends Migration
     public function up()
     {
         Schema::table('invoices', function (Blueprint $table) {
-            $table->smallInteger('payment_condition');
+            $table->integer('payment_condition');
         });
     }
